prayers_screen stub
